[google_sign_in] Android: Move GoogleSignInWrapper to a separate file. (#2636)
